chore: rename to `content-api`
